# api/openapi.yaml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  description: REST API to be used to create tableside food ordering applications.
  contact:
    name: Course Discord server
    url: https://discord.gg/dPsnz5u9
    email: mark.wentowski@docsgeek.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - url: http://localhost:80/
    description: Local setup
  - url: '{gitpod-server-url}/'
    description: Gitpod setup
    variables:
      gitpod-server-url:
        default: 'https://8080-<workspace-id>.gitpod.io'
        description: 'Dynamic Gitpod URL. Replace <workspace-id> with your Gitpod workspace ID if needed.'
...
paths:
  /orders:
    get:
      tags:
        - Orders
      summary: Get a list of orders
      description:
        This operation retrieves a list of orders based on specified query
        parameters. It is commonly used to view the `status` and `priority` of orders
        on the current day or view historical orders for inventory and analytics purposes.
      operationId: getOrderList
      parameters:
        - name: sort
          in: query
          description: |
            Specifies the field by which the results should be sorted. This parameter allows you to define the sorting order based on a particular field in the dataset. If you include a `sort` value, you must include a `order` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: priority
        - name: order
          in: query
          description: |
            Determines the order in which the results should be sorted. This parameter specifies whether the sorting should be in ascending or descending order. Valid values are `asc` for ascending order and `desc` for descending order. If you include an `order` value, you must include a `sort` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: asc
        - name: fields
          in: query
          description: |
            Use this parameter to specify which resource properties should be included in the response. Provide a list of schema property names, separated by the `|` delimiter. This allows for precise control over the data returned.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: table_number|status|priority
        - name: filter
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator~value`.

            **Basic Equality (eq)**
            - For exact matches, use `fieldName.eq~value`.

            **Comparison Operators:**
            - **Less Than (lt)** Use `fieldName.lt~value`.
            - **Less Than or Equal To (lte)** Use `fieldName.lte~value`.
            - **Greater Than (gt):** Use `fieldName.gt~value`.
            - **Greater Than or Equal To (gte):** Use `fieldName.gte~value`.

            Separate filters using the `|` delimter. Dates should be in `YYYY-MM-DD` format.
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: table_number.lt~5|status.eq~On the Way
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            type: integer
          example: 10
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Orders
    post:
      tags:
        - Orders
      summary: Create an order
      description: |
        This operation can be used to create an order when a patron places an order via a mobile app, tablet app, self-service kiosk, etc., depending on the API consumer's use case.
      operationId: postOrder
      requestBody:
        $ref: '#/components/requestBodies/CreateOrderRequest'
      responses:
        '201':
          description: |
            Resource created successfully. The response body contains the newly created resource.

            The server logic applies to the response body as follows:

            - The default `priority` is `3` out of `5`, indicating medium priority. Restaurant staff may adjust the priority based on factors like preparation time.
            - The `status` is set to `Received`. Restaurant staff update the status as the order progresses toward completion.
          headers:
            Location:
              description: Points to the resource endpoint for the newly created resource
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Orders
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get an order's details
      description: |
        This operation retrieves detailed information about a specific order identified by its unique order ID.
      operationId: getOrder
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bcc
        - name: fields
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: table_number|status|priority
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Orders
    put:
      tags:
        - Orders
      summary: Update an order
      description: |
        This operation updates an existing order with the provided data. You can modify the order's status, priority, and the list of dishes associated with it.
      operationId: putOrder
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bcd
      requestBody:
        $ref: '#/components/requestBodies/UpdateOrderRequest'
      responses:
        '200':
          description:
            Resource updated successfully. The response body contains the
            updated resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Orders
    delete:
      tags:
        - Orders
      summary: Delete an order
      description: |
        This operation deletes an existing order identified by its unique order ID.
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bca
      responses:
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Orders
  /orders/{id}/dishes:
    get:
      tags:
        - Orders
      summary: Retrieve dishes for an order
      description: |
        This operation retrieves a list of dishes belonging to a specific order. Each dish object returned contains detailed information about the dish and which ingredients are essential to preparing the dish.
      operationId: getOrderDishes
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bcb
        - name: sort
          in: query
          description: |
            Specifies the field by which the results should be sorted. This parameter allows you to define the sorting order based on a particular field in the dataset. If you include a `sort` value, you must include a `order` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: station
        - name: order
          in: query
          description: |
            Determines the order in which the results should be sorted. This parameter specifies whether the sorting should be in ascending or descending order. Valid values are `asc` for ascending order and `desc` for descending order.  If you include an `order` value, you must include a `sort` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: asc
        - name: fields
          in: query
          description: |
            Use this parameter to specify which resource properties should be included in the response. Provide a list of schema property names, separated by the `|` delimiter. This allows for precise control over the data returned.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name|station|ingredients
        - name: filter
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator~value`.

            **Basic Equality (eq)**
            - For exact matches, use `fieldName.eq~value`.

            **Comparison Operators:**
            - **Less Than (lt)** Use `fieldName.lt~value`.
            - **Less Than or Equal To (lte)** Use `fieldName.lte~value`.
            - **Greater Than (gt):** Use `fieldName.gt~value`.
            - **Greater Than or Equal To (gte):** Use `fieldName.gte~value`.

            Separate filters using the `|` delimter. Dates should be in `YYYY-MM-DD` format.
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: created_at.gt~2024-05-07|station.eq~hot
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            type: integer
          example: 10
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Orders
...
components:
  schemas:
...
    SpecialRequests:
      type: string
      description:
        "Any special instructions or additional requests associated with\
        \ the order, such as customization details or specific preferences from the\
        \ customer."
      nullable: true
...
    CreatedAt:
      type: string
      description: Timestamp when the resource was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
...
    DishIds:
      type: array
      description: Contains the unique identifiers of dishes to add to the order.
      example:
        - 692a4a34a46ecf001f35ac6f
        - 6a4a857c022eb7001f9b8972
        - 6b7c2441a0db43001f9bf377
      items:
        type: string
...
    Id:
      type: string
      description: The unique identifier for the resource.
      readOnly: true
      example: 674d0bf5c28b69001f8e03a1
...
    Order:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        name:
          pattern: "^[^\\s]*$"
          type: string
          description:
            The first name of the patron that placed the order. Cannot
            contain spaces.
          example: John
        table_number:
          $ref: '#/components/schemas/TableNumber'
        dish_ids:
          $ref: '#/components/schemas/DishIds'
        special_requests:
          $ref: '#/components/schemas/SpecialRequests'
        priority:
          $ref: '#/components/schemas/Priority'
        scheduled_at:
          $ref: '#/components/schemas/ScheduledAt'
        status:
          $ref: '#/components/schemas/Status'
      example:
        updated_at: 2024-04-22T10:00:00Z
        dish_ids:
          - 692a4a34a46ecf001f35ac6f
          - 6a4a857c022eb7001f9b8972
          - 6b7c2441a0db43001f9bf377
        name: John
        table_number: 8
        created_at: 2024-04-22T10:00:00Z
        scheduled_at: 2000-01-23T04:56:07.000+00:00
        id: 674d0bf5c28b69001f8e03a1
        special_requests: special_requests
        priority: 0
        status: Received
...
    Priority:
      maximum: 5
      minimum: 0
      type: integer
      description:
        "Order priority from `1` to `5` (i.e., greatest to least priority).\
        \ Priority is normally adjusted by kitchen staff."
      default: 3
    Problem:
      required:
        - detail
        - title
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
...
    ScheduledAt:
      type: string
      description: The timestamp indicating when the order is scheduled for.
      format: date-time
      nullable: true
...
    Status:
      type: string
      description:
        "The status of the order as it progresses toward completion:\n\
        - **`Draft`**: The patron has yet to submit their order.\n- **`Received`**:\
        \ The patron has placed an order, displaying details in the expeditor's KDS.\
        \ \n- **`In Progress`**: The expeditor prioritizes the order, displaying relevant\
        \ parts in the respective KDS stations for preparation.\n- **`Ready for assembly`**:\
        \ Stations mark dishes as ready, and the expeditor assembles the order.\n\
        \  - **Table-side**: The expeditor changes the status to `On the Way`, and\
        \ a food runner delivers it.\n  - **Takeaway**: The expeditor changes the\
        \ status to `Ready for pickup`, and the patron collects it at the pickup station.\n"
      nullable: true
      example: Received
      enum:
        - Draft
        - Received
        - In Progress
        - Ready for Assembly
        - On the Way
        - Ready for Pickup
    TableNumber:
      type: integer
      description: |
        The table number tells the staff where in the the patron is located. Set the value to `null` for a takeaway order.
      nullable: true
      example: 8
...
    UpdatedAt:
      type: string
      description: Timestamp when the order was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
...
    inline_response_200:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total_results:
          type: integer
          example: 20
      example:
        results:
          - updated_at: 2024-04-22T10:00:00Z
            dish_ids:
              - 692a4a34a46ecf001f35ac6f
              - 6a4a857c022eb7001f9b8972
              - 6b7c2441a0db43001f9bf377
            name: John
            table_number: 8
            created_at: 2024-04-22T10:00:00Z
            scheduled_at: 2000-01-23T04:56:07.000+00:00
            id: 674d0bf5c28b69001f8e03a1
            special_requests: special_requests
            priority: 0
            status: Received
          - updated_at: 2024-04-22T10:00:00Z
            dish_ids:
              - 692a4a34a46ecf001f35ac6f
              - 6a4a857c022eb7001f9b8972
              - 6b7c2441a0db43001f9bf377
            name: John
            table_number: 8
            created_at: 2024-04-22T10:00:00Z
            scheduled_at: 2000-01-23T04:56:07.000+00:00
            id: 674d0bf5c28b69001f8e03a1
            special_requests: special_requests
            priority: 0
            status: Received
        total_results: 20
    inline_response_200_1:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total_results:
          type: integer
          example: 1
      example:
        results:
          - updated_at: 2024-04-22T10:00:00Z
            dish_ids:
              - 692a4a34a46ecf001f35ac6f
              - 6a4a857c022eb7001f9b8972
              - 6b7c2441a0db43001f9bf377
            name: John
            table_number: 8
            created_at: 2024-04-22T10:00:00Z
            scheduled_at: 2000-01-23T04:56:07.000+00:00
            id: 674d0bf5c28b69001f8e03a1
            special_requests: special_requests
            priority: 0
            status: Received
          - updated_at: 2024-04-22T10:00:00Z
            dish_ids:
              - 692a4a34a46ecf001f35ac6f
              - 6a4a857c022eb7001f9b8972
              - 6b7c2441a0db43001f9bf377
            name: John
            table_number: 8
            created_at: 2024-04-22T10:00:00Z
            scheduled_at: 2000-01-23T04:56:07.000+00:00
            id: 674d0bf5c28b69001f8e03a1
            special_requests: special_requests
            priority: 0
            status: Received
        total_results: 1
    inline_response_200_2:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Dish'
        total_results:
          type: integer
          example: 1
      example:
        results:
          - image_name: burger
            price: null
            name: Risotto alla Milanese
            station: hot
            created_at: 2024-04-22T10:00:00Z
            description: A description of the resource.
            ingredients:
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
            preparation_time: 999
            id: 674d0bf5c28b69001f8e03a1
            category: Main Course
          - image_name: burger
            price: null
            name: Risotto alla Milanese
            station: hot
            created_at: 2024-04-22T10:00:00Z
            description: A description of the resource.
            ingredients:
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
            preparation_time: 999
            id: 674d0bf5c28b69001f8e03a1
            category: Main Course
        total_results: 1
...
  requestBodies:
...
    CreateOrderRequest:
      description: Request body containing information used to create the resource.
      content:
        application/json:
          schema:
            required:
              - dish_ids
              - name
            type: object
            properties:
              name:
                pattern: "^[^\\s]*$"
                type: string
                description:
                  The first name of the patron that placed the order. Cannot
                  contain spaces.
                example: John
              table_number:
                $ref: '#/components/schemas/TableNumber'
              dish_ids:
                $ref: '#/components/schemas/DishIds'
              special_requests:
                $ref: '#/components/schemas/SpecialRequests'
              scheduled_at:
                $ref: '#/components/schemas/ScheduledAt'
          examples:
            Order - Unscheduled:
              value:
                name: Ben
                table_number: 18
                dish_ids:
                  - 692a4a34a46ecf001f35ac6f
                  - 6a4a857c022eb7001f9b8972
                  - 6b7c2441a0db43001f9bf377
                special_requests: No peanuts.
            Order - Scheduled:
              value:
                name: Ben
                table_number: 18
                dish_ids:
                  - 692a4a34a46ecf001f35ac6f
                  - 6a4a857c022eb7001f9b8972
                  - 6b7c2441a0db43001f9bf377
                scheduled_at: 2024-08-26T09:03:18.865Z
      required: true
    UpdateOrderRequest:
      description: Request body containing properties to update the resource.
      content:
        application/json:
          schema:
            type: object
            properties:
              dish_ids:
                $ref: '#/components/schemas/DishIds'
              priority:
                $ref: '#/components/schemas/Priority'
              status:
                $ref: '#/components/schemas/Status'
      required: true
...
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.
      scheme: basic